sco_repack: Try to fix reading/parsing Wavefront OBJ files from Wings3D. Reported by kraggrim#0650 on Discord.
The lines look like f 1041// 1072// 1040//, so my reader thinks the // are comments and strips anything to the right. Usually these lines should look like  f 1041/3/5 1072/6/3 1040/5/3, or f 1041 1072 1040.

Stop stripping // comments, plus, fix stripping # comments that start right at the beginning of the line. `return pos != -1 and line[:pos] or line` failed because line[:pos] was '', which is False.